move IssueDateTime from Minimum to DataType
