+++ src/DataType/IssueDateTime.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tiime\CrossIndustryInvoice\Minimum;
+namespace Tiime\CrossIndustryInvoice\DataType;
 
 class IssueDateTime
 {
